websocket/pyproject.toml
[tool.poetry]
name = "starters-ws"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-decouple = "^3.8"
fastapi = "^0.115.6"
pyjwt = "^2.10.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
